Rename lib/ttnt/tasks.rb to lib/ttnt/task_helper.rb
