+++ Rakefile
@@ -1,6 +1,6 @@
 require "bundler/gem_tasks"
 require "rake/testtask"
-require "ttnt/tasks"
+require "ttnt/testtask"
 
 Rake::TestTask.new(:test) do |t|
   t.libs << "test"
